--- Kodukov/lab3/T03BST/T03BST.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "T03BST", "T03BST.vcxproj", "{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Debug|x64.ActiveCfg = Debug|x64
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Debug|x64.Build.0 = Debug|x64
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Debug|x86.ActiveCfg = Debug|Win32
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Debug|x86.Build.0 = Debug|Win32
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Release|x64.ActiveCfg = Release|x64
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Release|x64.Build.0 = Release|x64
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Release|x86.ActiveCfg = Release|Win32
-		{6D33508B-3CD6-476B-B2AC-D8A5FADEE2CF}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {867CBB5C-E74E-412A-9DBA-D5C30AC55366}
-	EndGlobalSection
-EndGlobal
