+++ Kodukov/lab3/Tests/input.txt
@@ -1,0 +1,11 @@
+1,/2,/3,/4,/5,/////
+1,2,3,4,5,//////////
+0,0,//0,//
+1,//
+10,9,8,/7,///6,5,//4,3,//2,1,///
+6,5,3,/4,///8,7,//12,11,//15,13,///	
+
+
+
+
+
